moving logic into servoce type objects
